fix: Process Shift+Enter as pure Enter
Before this commit, Shift+Enter adds a new line in the same block, and
result in a wrong line number. It is somehow fine if only the line
number is wrong, but with the language server, the squiggles are also
in the wrong place, thus it's necessary to fix this.